refactor: refactor App component by creating FormContainer and AnticipatedAmountsContainer components
