+++ src/components/FormContainer/styles.ts
@@ -1,0 +1,7 @@
+import { CSSProperties } from "react";
+
+export const override: CSSProperties = {
+  display: "block",
+  margin: "0 auto",
+  borderColor: "red",
+};
