+++ src/components/ThemeButton/styles.ts
@@ -12,6 +12,8 @@ export const StyledButton = styled.button`
 
   cursor: pointer;
 
+  height: 2rem;
+
   transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
 
   &:hover {
